+++ move_selection.h
@@ -47,7 +47,7 @@ typedef struct {
     int num_killers;
     int moves_so_far;
     int quiet_moves_so_far;
-    int depth;
+    float depth;
     position_t* pos;
     bool single_reply;
 } move_selector_t;
